// apps/frontend/src/app/proposals/components/components/types.ts
export enum ProposalFilter {
  State = "State",
  InThisRound = "In this round",
  LookingForSupport = "Looking for support",
  UpcomingVoting = "Upcoming voting",
}

export enum StateFilter {
  Active = "Active",
  Canceled = "Canceled",
  Defeated = "Defeated",
  Succeeded = "Succeeded",
  Queued = "Queued",
  Executed = "Executed",
  DepositNotMet = "Deposit not met",
}